server/requirements.txt
flask
flask_cors
python-dotenv
google-generativeai
pandas
python-dotenv